Changed DB to MySQL, and minor changes in entity classes

+++ src/main/java/com/htadg/taskit/repository/BoardRepository.java
@@ -4,9 +4,7 @@ import com.htadg.taskit.entity.Board;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.util.Collection;
-
 @Repository
-public interface BoardRepository extends JpaRepository<Board, Integer> {
+public interface BoardRepository extends JpaRepository<Board, Long> {
     public Board findByName(String name);
 }
